refactor: enable platform interface on Android
- Performance Attributes are extended by
  Trace and HTTP Metric.
- Creation of instance on Android side assumes
  instance is created by core.
- HTTP Metric and Trace are split into separate
  files.
