+++ packages/firebase_performance/firebase_performance/lib/firebase_performance.dart
@@ -7,8 +7,12 @@ library firebase_performance;
 
 import 'dart:async';
 
-import 'package:flutter/services.dart';
-import 'package:flutter/foundation.dart';
+import 'package:firebase_core/firebase_core.dart';
+import 'package:firebase_core_platform_interface/firebase_core_platform_interface.dart';
+import 'package:firebase_performance_platform_interface/firebase_performance_platform_interface.dart';
+
+export 'package:firebase_performance_platform_interface/firebase_performance_platform_interface.dart'
+    show HttpMethod;
 
 part 'src/firebase_performance.dart';
 part 'src/http_metric.dart';
