+++ packages/firebase_performance/firebase_performance_platform_interface/lib/firebase_performance_platform_interface.dart
@@ -1,0 +1,6 @@
+library firebase_performance_platform_interface;
+
+export 'src/platform_interface/platform_interface_attributes.dart';
+export 'src/platform_interface/platform_interface_firebase_performance.dart';
+export 'src/platform_interface/platform_interface_http_metric.dart';
+export 'src/platform_interface/platform_interface_trace.dart';
